api_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pika==1.3.2
requests==2.31.0
Pillow==10.1.0
